% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{add_context_name}
\alias{add_context_name}
\title{Add context names to tasks information}
\usage{
add_context_name(tasks, contexts)
}
\arguments{
\item{tasks}{data frame with tidy tasks information}

\item{contexts}{data frame with tidy context information}
}
\value{
data frame of task information with named context
}
\description{
Task information is stored with an associated context ID
However for analysis context name is far more useful
This function takes context and task data frames and converts
context ID to context name.
}

