% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{add_project_name}
\alias{add_project_name}
\title{Add project names to tasks information}
\usage{
add_project_name(tasks, projects)
}
\arguments{
\item{tasks}{data frame with tidy tasks information}

\item{projects}{data frame with tidy project information}
}
\value{
data frame of tasks with project name added
}
\description{
Add project names to tasks information
}
