% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_contexts}
\alias{filter_contexts}
\title{This function filters for all context nodes}
\usage{
filter_contexts(x)
}
\arguments{
\item{x}{input document}
}
\value{
nodeset of all contexts
}
\description{
This function filters for all context nodes
}
