% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_projects}
\alias{filter_projects}
\title{This function filters for all project-like tasks}
\usage{
filter_projects(x)
}
\arguments{
\item{x}{input document}
}
\value{
nodeset of all projects
}
\description{
This function filters for all project-like tasks
}
