% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_task_to_tag}
\alias{filter_task_to_tag}
\title{This function filters for all task-to-tag nodes}
\usage{
filter_task_to_tag(x)
}
\arguments{
\item{x}{input document}
}
\value{
nodeset of all task-to-tag
}
\description{
This function filters for all task-to-tag nodes
}
