% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_tasks}
\alias{filter_tasks}
\title{This function filters for all tasks}
\usage{
filter_tasks(x)
}
\arguments{
\item{x}{input document}
}
\value{
nodeset of all tasks
}
\description{
This function filters for all tasks
}
