% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{is_flagged}
\alias{is_flagged}
\title{Converts the xml formatting of flagged data into an R logical object}
\usage{
is_flagged(x)
}
\arguments{
\item{x}{a string}
}
\value{
logical that define if a task/project was flagged
}
\description{
Converts the xml formatting of flagged data into an R logical object
}
