% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_context_node}
\alias{parse_context_node}
\title{Parse the data from a single context node}
\usage{
parse_context_node(node, ns)
}
\arguments{
\item{node}{the node for a single context}

\item{ns}{xml namespace object}
}
\value{
list containing the data of a single context
}
\description{
Parse the data from a single context node
}
