% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeset-processing.R
\name{parse_contexts}
\alias{parse_contexts}
\title{Retrieve contexts from OmniFocus database file}
\usage{
parse_contexts(x)
}
\arguments{
\item{x}{the XML document in xml2 document format}
}
\value{
data frame in raw format containing context information
}
\description{
Retrieve contexts from OmniFocus database file
}
