% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_project_node}
\alias{parse_project_node}
\title{Parse the data from a single project node}
\usage{
parse_project_node(node, ns)
}
\arguments{
\item{node}{the node for a single task}

\item{ns}{xml namespace object}
}
\value{
list containing the data of a single project
}
\description{
Parse the data from a single project node
}
