% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-parsing.R
\name{parse_projects}
\alias{parse_projects}
\title{Retrieve projects from OmniFocus database file}
\usage{
parse_projects(x)
}
\arguments{
\item{x}{the XML document in xml2 document format}
}
\value{
data frame containing project information
}
\description{
Retrieve projects from OmniFocus database file
}

