% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_task_node}
\alias{parse_task_node}
\title{Parse a task node}
\usage{
parse_task_node(node, ns)
}
\arguments{
\item{node}{the task node}

\item{ns}{xml namespace object}
}
\value{
list containing the data of a single task
}
\description{
Parse a task node
}
