% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_task_to_tag_node}
\alias{parse_task_to_tag_node}
\title{Parse the data from a single task-to-tag node}
\usage{
parse_task_to_tag_node(node, ns)
}
\arguments{
\item{node}{the node for a single task-to-tag}

\item{ns}{xml namespace object}
}
\value{
list containing the data of a single task-to-tag
}
\description{
Parse the data from a single task-to-tag node
}
