% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeset-processing.R
\name{parse_task_to_tags}
\alias{parse_task_to_tags}
\title{Retrieve task-to-tag nodes from OmniFocus database file}
\usage{
parse_task_to_tags(x)
}
\arguments{
\item{x}{the XML document in xml2 document format}
}
\value{
data frame in raw format containing task-to-tag associations
}
\description{
Retrieve task-to-tag nodes from OmniFocus database file
}
