% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeset-processing.R
\name{parse_tasks}
\alias{parse_tasks}
\title{Retrieve tasks from OmniFocus database file}
\usage{
parse_tasks(x)
}
\arguments{
\item{x}{the XML document in xml2 document format}
}
\value{
data frame in raw format containing task information
}
\description{
Retrieve tasks from OmniFocus database file
}
