% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_db.R
\name{process_of_db}
\alias{process_of_db}
\title{Process an OmniFocus content XML file}
\usage{
process_of_db(path)
}
\arguments{
\item{path}{full path to OmniFocus XML file}
}
\value{
data frame containing task information in tidy format
}
\description{
Process an OmniFocus content XML file
}
\examples{
\dontrun{
of_data <- process_of_db("~/Downloads/OmniFocus/contents.xml")
}
}
