% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_projects}
\alias{tidy_projects}
\title{Tidy project data retrieved from OmniFocus database file}
\usage{
tidy_projects(.data)
}
\arguments{
\item{.data}{the data frame constructed by parse_projects}
}
\value{
data frame containing project information in tidy format
}
\description{
Tidy project data retrieved from OmniFocus database file
}
