% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-parsing.R
\name{tidy_tasks}
\alias{tidy_tasks}
\title{Tidy task data retrieved from OmniFocus database file}
\usage{
tidy_tasks(.data)
}
\arguments{
\item{.data}{the data frame constructed by parse_tasks}
}
\value{
data frame containing task information in tidy format
}
\description{
Tidy task data retrieved from OmniFocus database file
}

