% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-parsing.R
\name{parse_contexts}
\alias{parse_contexts}
\title{Retrieve contexts from OmniFocus database file}
\usage{
parse_contexts(x)
}
\arguments{
\item{x}{an xml2 document built from the contents.xml}
}
\value{
data frame containing context name and internal ID
}
\description{
Retrieve contexts from OmniFocus database file
}
