% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeset-processing.R
\name{parse_items}
\alias{parse_items}
\title{Retrieve projects from OmniFocus database file}
\usage{
parse_items(x, filter, parser)
}
\arguments{
\item{x}{the XML document in xml2 document format}

\item{filter}{a function that takes an xml_document and returns an xml_nodeset, used to filter for xml tags}

\item{parser}{a function which takes an xml_node and returns a list, used to parse an xml node into a list}
}
\value{
data frame containing project information
}
\description{
Retrieve projects from OmniFocus database file
}
